优化显示

+++ backend/controllers/article_controller.go
@@ -405,6 +405,16 @@ func (c *ArticleController) DeleteArticle(ctx *gin.Context) {
 		ctx.JSON(http.StatusOK, utils.Error(1, "文章ID无效"))
 		return
 	}
+	userID := ctx.GetUint("userID")
+	article, err := c.articleService.GetArticleByID(uint(id))
+	if err != nil {
+		ctx.JSON(http.StatusOK, utils.Error(1, "文章不存在"))
+		return
+	}
+	if strconv.FormatUint(uint64(userID), 10) != article.AuthorCode {
+		ctx.JSON(http.StatusOK, utils.Error(1, "无权操作"))
+		return
+	}
 	err = c.articleService.DeleteArticle(uint(id))
 	if err != nil {
 		ctx.JSON(http.StatusOK, utils.Error(1, "删除文章失败: "+err.Error()))
